requirements.txt
asgiref==3.11.0
awaiter==0.1.2
Django==6.0.1
django-cors-headers==4.9.0
djangorestframework==3.16.1
gunicorn==25.1.0
whitenoise==6.11.0
mysqlite==0.0.0
packaging==26.0
psycopg2-binary==2.9.11
PyMySQL==1.1.2
sqlparse==0.5.5
tzdata==2025.3
